% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hull_metrics}
\alias{hull_metrics}
\title{Calculate crown attributes for a set of tree points}
\usage{
hull_metrics(X, Y, Z, alpha = 1)
}
\arguments{
\item{X}{Numeric vector, x-coordinates of tree points}

\item{Y}{Numeric vector, y-coordinates of tree points}

\item{Z}{Numeric vector, z-coordinates of tree points}
}
\value{
A data frame with columns for each crown attribute.
}
\description{
This function calculates various crown attributes for a set of 3D tree points including:
crown height (Zmax, Zq999, Zq99, Z_mean), crown size (n_points, vol_convex, vol_concave),
and crown complexity (CV_Z, CRR).
}
